+++ composer.json
@@ -91,10 +91,8 @@
             "EonX\\EasyActivity\\": "packages/EasyActivity/src",
             "EonX\\EasyApiPlatform\\": "packages/EasyApiPlatform/src",
             "EonX\\EasyApiToken\\": "packages/EasyApiToken/src",
-            "EonX\\EasyAsync\\": [
-                "packages/EasyAsync/bundle",
-                "packages/EasyAsync/src"
-            ],
+            "EonX\\EasyAsync\\": "packages/EasyAsync/src",
+            "EonX\\EasyAsync\\Bundle\\": "packages/EasyAsync/bundle",
             "EonX\\EasyAsync\\Laravel\\": "packages/EasyAsync/laravel",
             "EonX\\EasyBankFiles\\": "packages/EasyBankFiles/src",
             "EonX\\EasyBatch\\": "packages/EasyBatch/src",
