+++ packages/EasyAsync/bundle/config/doctrine_persistent_connections.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
+use EonX\EasyAsync\Bundle\Enum\ConfigParam;
 use EonX\EasyAsync\Doctrine\Closer\ManagersCloser;
-use EonX\EasyAsync\Enum\ConfigParam;
 use EonX\EasyAsync\Messenger\Listener\ClosePersistentConnectionListener;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
