+++ packages/EasyAsync/composer.json
@@ -6,6 +6,7 @@
     "require": {
         "php": "^8.1",
         "doctrine/orm": "^2.14",
+        "doctrine/dbal": "^3.0",
         "doctrine/persistence": "^3.0",
         "psr/log": "^3.0",
         "symfony/clock": "^6.4",
@@ -21,8 +22,6 @@
         "symfony/var-exporter": "^6.4"
     },
     "require-dev": {
-        "doctrine/common": "^3.0",
-        "doctrine/dbal": "^3.0",
         "eonx-com/easy-event-dispatcher": "^5.11",
         "eonx-com/easy-random": "^5.11",
         "mockery/mockery": "^1.6",
@@ -31,10 +30,8 @@
     },
     "autoload": {
         "psr-4": {
-            "EonX\\EasyAsync\\": [
-                "bundle",
-                "src"
-            ],
+            "EonX\\EasyAsync\\": "src",
+            "EonX\\EasyAsync\\Bundle\\": "bundle",
             "EonX\\EasyAsync\\Laravel\\": "laravel"
         }
     },
@@ -44,7 +41,6 @@
         }
     },
     "suggest": {
-        "doctrine/dbal": "To use Doctrine related middlewares",
         "eonx-com/easy-error-handler": "To report about the Illuminate worker stopping event",
         "eonx-com/easy-logging": "To provide the logging service",
         "illuminate/contracts": "To use this package with Laravel/Lumen",
