+++ packages/EasyAsync/laravel/Listeners/DoctrineManagersClearListener.php
@@ -21,8 +21,8 @@ final class DoctrineManagersClearListener
     }
 
     /**
-     * @throws \EonX\EasyAsync\Doctrine\Exceptions\DoctrineConnectionNotOkException
-     * @throws \EonX\EasyAsync\Doctrine\Exceptions\DoctrineManagerClosedException
+     * @throws \EonX\EasyAsync\Doctrine\Exception\DoctrineConnectionNotOkException
+     * @throws \EonX\EasyAsync\Doctrine\Exception\DoctrineManagerClosedException
      */
     public function handle(JobProcessing $event): void
     {
