+++ packages/EasyAsync/src/Doctrine/Exception/DoctrineManagerClosedException.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyAsync\Doctrine\Exceptions;
+namespace EonX\EasyAsync\Doctrine\Exception;
 
 use EonX\EasyAsync\Common\Exception\AbstractEasyAsyncException;
 use EonX\EasyAsync\Common\Exception\ShouldKillWorkerExceptionInterface;
 
-final class DoctrineConnectionNotOkException extends AbstractEasyAsyncException implements
+final class DoctrineManagerClosedException extends AbstractEasyAsyncException implements
     ShouldKillWorkerExceptionInterface
 {
     // No body needed
