+++ packages/EasyAsync/tests/Doctrine/Checker/ManagersSanityCheckerTest.php
@@ -7,8 +7,8 @@ use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
 use Doctrine\Persistence\Mapping\Driver\PHPDriver;
 use EonX\EasyAsync\Doctrine\Checker\ManagersSanityChecker;
-use EonX\EasyAsync\Doctrine\Exceptions\DoctrineConnectionNotOkException;
-use EonX\EasyAsync\Doctrine\Exceptions\DoctrineManagerClosedException;
+use EonX\EasyAsync\Doctrine\Exception\DoctrineConnectionNotOkException;
+use EonX\EasyAsync\Doctrine\Exception\DoctrineManagerClosedException;
 use EonX\EasyAsync\Tests\AbstractStoreTestCase;
 use EonX\EasyAsync\Tests\Stub\ConnectionStub;
 use EonX\EasyAsync\Tests\Stub\EntityManagerForSanityStub;
